move db logic into separate file
