+++ stac_fastapi/elasticsearch/stac_fastapi/elasticsearch/core.py
@@ -19,6 +19,7 @@ from stac_pydantic.shared import MimeTypes
 from stac_fastapi.elasticsearch import serializers
 from stac_fastapi.elasticsearch.config import ElasticsearchSettings
 from stac_fastapi.elasticsearch.session import Session
+from stac_fastapi.elasticsearch.database_logic import CoreDatabaseLogic
 
 # from stac_fastapi.elasticsearch.types.error_checks import ErrorChecks
 from stac_fastapi.types.core import BaseCoreClient
@@ -31,208 +32,6 @@ NumType = Union[float, int]
 
 ITEMS_INDEX = "stac_items"
 COLLECTIONS_INDEX = "stac_collections"
-
-
-def mk_item_id(item_id: str, collection_id: str):
-    """Make the Elasticsearch document _id value from the Item id and collection."""
-    return f"{item_id}|{collection_id}"
-
-@attr.s
-class DatabaseLogic():
-    """Core database logic."""
-    settings = ElasticsearchSettings()
-    client = settings.create_client
-    item_serializer: Type[serializers.Serializer] = attr.ib(
-        default=serializers.ItemSerializer
-    )
-    collection_serializer: Type[serializers.Serializer] = attr.ib(
-        default=serializers.CollectionSerializer
-    )
-
-    @staticmethod
-    def bbox2poly(b0, b1, b2, b3):
-        """Transform bbox to polygon."""
-        poly = [[[b0, b1], [b2, b1], [b2, b3], [b0, b3], [b0, b1]]]
-        return poly
-
-    def get_all_collections(self, base_url: str) -> List:
-        """Database logic to retrieve a list of all collections."""
-        try:
-            collections = self.client.search(
-                index=COLLECTIONS_INDEX, query={"match_all": {}}
-            )
-        except elasticsearch.exceptions.NotFoundError:
-            raise NotFoundError("No collections exist")
-
-        serialized_collections = [
-            self.collection_serializer.db_to_stac(
-                collection["_source"], base_url=base_url
-            )
-            for collection in collections["hits"]["hits"]
-        ]
-
-        return serialized_collections
-
-    def get_one_collection(self, collection_id: str):
-        try:
-            collection = self.client.get(index=COLLECTIONS_INDEX, id=collection_id)
-        except elasticsearch.exceptions.NotFoundError:
-            raise NotFoundError(f"Collection {collection_id} not found")
-
-        return collection["_source"]
-
-    def get_item_collection(self, collection_id: str, limit: int, base_url: str):
-        search = Search(using=self.client, index="stac_items")
-
-        collection_filter = Q(
-            "bool", should=[Q("match_phrase", **{"collection": collection_id})]
-        )
-        search = search.query(collection_filter)
-        try:
-            count = search.count()
-        except elasticsearch.exceptions.NotFoundError:
-            raise NotFoundError("No items exist")
-        # search = search.sort({"id.keyword" : {"order" : "asc"}})
-        search = search.query()[0:limit]
-        collection_children = search.execute().to_dict()
-
-        serialized_children = [
-            self.item_serializer.db_to_stac(item["_source"], base_url=base_url)
-            for item in collection_children["hits"]["hits"]
-        ]
-
-        return serialized_children, count
-
-    def get_one_item(self, collection_id: str, item_id: str):
-        try:
-            item = self.client.get(
-                index=ITEMS_INDEX, id=mk_item_id(item_id, collection_id)
-            )
-        except elasticsearch.exceptions.NotFoundError:
-            raise NotFoundError(
-                f"Item {item_id} does not exist in Collection {collection_id}"
-            )
-        return item["_source"]
-
-    def create_search_object(self):
-        search = (
-            Search()
-            .using(self.client)
-            .index(ITEMS_INDEX)
-            .sort(
-                {"properties.datetime": {"order": "desc"}},
-                {"id": {"order": "desc"}},
-                {"collection": {"order": "desc"}},
-            )
-        )
-
-        return search
-
-    def create_query_filter(self, search, op: str, field: str, value: float):
-        if op != "eq":
-            key_filter = {field: {f"{op}": value}}
-            search = search.query(Q("range", **key_filter))
-        else:
-            search = search.query("match_phrase", **{field: value})
-
-        return search
-
-    def search_ids(self, search, item_ids: List):
-        id_list = []   
-        for item_id in item_ids:
-            id_list.append(Q("match_phrase", **{"id": item_id}))
-        id_filter = Q("bool", should=id_list)
-        search = search.query(id_filter)
-
-        return search
-
-    def search_collections(self, search, collection_ids: List):
-        collection_list = []
-        for collection_id in collection_ids:
-            collection_list.append(
-                Q("match_phrase", **{"collection": collection_id})
-            )
-        collection_filter = Q("bool", should=collection_list)
-        search = search.query(collection_filter)
-
-        return search
-
-    def search_datetime(self, search, datetime_search):
-        if "eq" in datetime_search:
-            search = search.query(
-                "match_phrase", **{"properties__datetime": datetime_search["eq"]}
-            )
-        else:
-            search = search.filter(
-                "range", properties__datetime={"lte": datetime_search["lte"]}
-            )
-            search = search.filter(
-                "range", properties__datetime={"gte": datetime_search["gte"]}
-            )
-        return search
-
-    def search_bbox(self, search, bbox: List):
-        poly = self.bbox2poly(bbox[0], bbox[1], bbox[2], bbox[3])
-        bbox_filter = Q(
-            {
-                "geo_shape": {
-                    "geometry": {
-                        "shape": {"type": "polygon", "coordinates": poly},
-                        "relation": "intersects",
-                    }
-                }
-            }
-        )
-        search = search.query(bbox_filter)
-        return search
-
-    def search_intersects(self, search, intersects: dict):
-        """Database logic to search a geojson object."""
-        intersect_filter = Q(
-            {
-                "geo_shape": {
-                    "geometry": {
-                        "shape": {
-                            "type": intersects.type.lower(),
-                            "coordinates": intersects.coordinates,
-                        },
-                        "relation": "intersects",
-                    }
-                }
-            }
-        )
-        search = search.query(intersect_filter)
-        return search
-
-    def sort_field(self, search, field, direction):
-        search = search.sort({field: {"order": direction}})
-        return search
-
-    def search_count(self, search):
-        try:
-            count = search.count()
-        except elasticsearch.exceptions.NotFoundError:
-            raise NotFoundError("No items exist")
-        
-        return count
-
-    def execute_search(self, search, limit: int, base_url: str) -> dict:
-        """Database logic to execute search with limit."""
-        search = search.query()[0 : limit]
-        response = search.execute().to_dict()
-
-        if len(response["hits"]["hits"]) > 0:
-            response_features = [
-                self.item_serializer.db_to_stac(item["_source"], base_url=base_url)
-                for item in response["hits"]["hits"]
-            ]
-        else:
-            response_features = []
-
-        return response_features
-        
-        
-
     
 @attr.s
 class CoreCrudClient(BaseCoreClient):
@@ -247,7 +46,7 @@ class CoreCrudClient(BaseCoreClient):
     )
     settings = ElasticsearchSettings()
     client = settings.create_client
-    database = DatabaseLogic()
+    database = CoreDatabaseLogic()
 
     @overrides
     def all_collections(self, **kwargs) -> Collections:
@@ -401,16 +200,6 @@ class CoreCrudClient(BaseCoreClient):
     def post_search(self, search_request: Search, **kwargs) -> ItemCollection:
         """POST search catalog."""
         base_url = str(kwargs["request"].base_url)
-        # search = (
-        #     Search()
-        #     .using(self.client)
-        #     .index(ITEMS_INDEX)
-        #     .sort(
-        #         {"properties.datetime": {"order": "desc"}},
-        #         {"id": {"order": "desc"}},
-        #         {"collection": {"order": "desc"}},
-        #     )
-        # )
         search = self.database.create_search_object()
 
         if search_request.query:
@@ -419,51 +208,17 @@ class CoreCrudClient(BaseCoreClient):
             for (field_name, expr) in search_request.query.items():
                 field = "properties__" + field_name
                 for (op, value) in expr.items():
-                    # if op != "eq":
-                    #     key_filter = {field: {f"{op}": value}}
-                    #     search = search.query(Q("range", **key_filter))
-                    # else:
-                    #     search = search.query("match_phrase", **{field: value})
-
                     search = self.database.create_query_filter(search=search, op=op, field=field, value=value)
 
         if search_request.ids:
             search = self.database.search_ids(search=search, item_ids=search_request.ids)
-            # id_list = []
-            
-            # for item_id in search_request.ids:
-            #     id_list.append(Q("match_phrase", **{"id": item_id}))
-            # id_filter = Q("bool", should=id_list)
-            # search = search.query(id_filter)
-
-
 
         if search_request.collections:
             search = self.database.search_ids(search_request.collections)
-            # collection_list = []
-            # for collection_id in search_request.collections:
-            #     collection_list.append(
-            #         Q("match_phrase", **{"collection": collection_id})
-            #     )
-            # collection_filter = Q("bool", should=collection_list)
-            # search = search.query(collection_filter)
 
         if search_request.datetime:
             datetime_search = self._return_date(search_request.datetime)
             search = self.database.search_datetime(datetime_search)
-
-
-            # if "eq" in datetime_search:
-            #     search = search.query(
-            #         "match_phrase", **{"properties__datetime": datetime_search["eq"]}
-            #     )
-            # else:
-            #     search = search.filter(
-            #         "range", properties__datetime={"lte": datetime_search["lte"]}
-            #     )
-            #     search = search.filter(
-            #         "range", properties__datetime={"gte": datetime_search["gte"]}
-            #     )
 
         if search_request.bbox:
             bbox = search_request.bbox
@@ -471,18 +226,6 @@ class CoreCrudClient(BaseCoreClient):
                 bbox = [bbox[0], bbox[1], bbox[3], bbox[4]]
 
             search = self.database.search_bbox(bbox=bbox)
-
-            # bbox_filter = Q(
-            #     {
-            #         "geo_shape": {
-            #             "geometry": {
-            #                 "shape": {"type": "polygon", "coordinates": poly},
-            #                 "relation": "intersects",
-            #             }
-            #         }
-            #     }
-            # )
-            # search = search.query(bbox_filter)
 
         if search_request.intersects:
             self.database.search_intersects(search_request.intersects)
@@ -493,7 +236,6 @@ class CoreCrudClient(BaseCoreClient):
                     sort.field = "properties__datetime"
                 field = sort.field + ".keyword"
                 search = self.database.sort_field(search=search, field=field, direction=sort.direction)
-                # search = search.sort({field: {"order": sort.direction}})
 
         count = self.database.search_count(search=search)
